SwiftTailCC: teach verifier musttail rules applicable to this CC.
SwiftTailCC has a different set of requirements than the C calling
convention for a tail call. The exact argument sequence doesn't have to
match, but fewer ABI-affecting attributes are allowed (mostly to be
conservative and because we have a reasonable idea of what might
actually be useful).

Also make sure the musttail diagnostic triggers if a musttail call isn't
actually a tail call.